// frontend/ui/src/assets/balkanapp/familytree.d.ts
declare class FamilyTree extends FamilyTreeBase {
// ...
  ): Array<{
    id: number | string;
    name: string;
    __score: number;
    __searchField: string;
    __searchMarks: string;
  }>;
// ...
}

declare namespace FamilyTree {
// ...
  interface options {
// ...
    state?: {
      name?: string;
      readFromLocalStorage?: boolean;
      writeToLocalStorage?: boolean;
      readFromIndexedDB?: boolean;
      writeToIndexedDB?: boolean;
      readFromUrlParams?: boolean;
      writeToUrlParams?: boolean;
    };
// ...
  }
// ...
}